Rename blockRendererUtils to blocksRendererUtils
